refactor: reorganizar componentes de edição de perfil e atualizar validação de telefone.
